fixed genericuser and uuid stuff
fixed genericuser and uuid stuff, and modules
