+++ src/App.vue
@@ -5,7 +5,8 @@
 </template>
 
 <script>
-export default {};
+export default {
+};
 </script>
 
 <style lang="scss">
