+++ src/store/modules/menu.js
@@ -1,0 +1,12 @@
+export default {
+  state: {
+    showMenu: false
+  },
+  getters: {},
+  mutations: {
+    showMenu(state) {
+      state.showMenu = !state.showMenu;
+    }
+  },
+  actions: {}
+};
